requirements.txt
google-genai>=1.11.0
python-dotenv>=1.0.1
